refactor(core): update order page to use slug
